feat: useFetch in general settings is removed and moved to use query
